pulse: remove redundant typecasts

+++ pulse.c
@@ -542,8 +542,8 @@ static int op_pulse_mixer_get_volume(int *l, int *r)
 								     NULL));
 	}
 
-	*l = (int)pa_cvolume_get_position(&pa_vol, &pa_cmap, PA_CHANNEL_POSITION_FRONT_LEFT);
-	*r = (int)pa_cvolume_get_position(&pa_vol, &pa_cmap, PA_CHANNEL_POSITION_FRONT_RIGHT);
+	*l = pa_cvolume_get_position(&pa_vol, &pa_cmap, PA_CHANNEL_POSITION_FRONT_LEFT);
+	*r = pa_cvolume_get_position(&pa_vol, &pa_cmap, PA_CHANNEL_POSITION_FRONT_RIGHT);
 
 	return rc;
 }
